// packages/aws-info/src/types.ts
// Generated by "npm run generate_info" in tradle/aws-utils repository
export type RegionCode = 'us-east-1' | 'us-east-2' | 'us-west-1' | 'us-west-2' | 'ca-central-1' | 'ap-south-1' | 'ap-northeast-1' | 'ap-northeast-2' | 'ap-northeast-3' | 'ap-southeast-1' | 'ap-southeast-2' | 'cn-north-1' | 'cn-northwest-1' | 'eu-central-1' | 'eu-west-1' | 'eu-west-2' | 'eu-west-3' | 'eu-north-1' | 'ap-east-1' | 'sa-east-1' | 'me-south-1' | 'ap-southeast-3' | 'us-gov-west-1' | 'eu-south-1' | 'us-gov-east-1' | 'af-south-1'
export type RegionIdx = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | 'a' | 'b' | 'c' | 'd' | 'e' | 'f' | 'g' | 'h' | 'i' | 'j' | 'k' | 'l' | 'm' | 'n' | 'o' | 'p'

export interface Coordinates {
  latitude: number
  longitude: number
}

export interface OnMap {
  loc: Coordinates
}

export enum Partition {
  Default = '*',
  UsGov = 'us-gov',
  China = 'cn'
}

export interface Region extends Partial<OnMap> {
  /**
   * Code as used in AWS
   */
  code: RegionCode

  /**
    * Short code used for suffix alike.
    * Fixed to never change
    */
  idx: RegionIdx

  /**
   * An AWS Region is always part of a group aka. partition
   */
  partition: Partition

  /**
   * Nearest fallback regions
   */
  fallback?: {
    readonly [type in 'ses' | 'sns']: RegionCode
  }
}

export enum SenderID {
  Unavailable = 0,
  Available = 1,
  PreRequisitesRequired = 2
}

export interface CountrySMS {
  readonly senderID: SenderID
  readonly twoWaySMS: boolean
}

export interface Country extends OnMap {
  id: string
  cca3: string
  callingCodes: readonly string[]
  title: string
  sms?: CountrySMS
}